// src/models/lesson.ts
export interface ILesson {
  name: string;
  lesson_fb: string;
  stage: number;
  type: string;
  weight: boolean;
  weighing: number;
  description: string;
  html: object;
  message: string;
  activity_id: string;
  assign: number;
  claps: number;
  client_id: string;
  competencies_json: object;
  course_fb: string;
  created_at: string;
  created_by: string;
  creating: string;
  deleted_at: string;
  embed_json: object;
  eval_attempts: number;
  eval_question_to_evaluate: number;
  hide: boolean;
  hours: number;
  image_url: string;
  index: number;
  is_deleted: boolean;
  is_individual: boolean;
  is_post: boolean;
  lecture: object;
  minutes: number;
  module_id: string;
  privacy: boolean;
  random: boolean;
  resources_json: object;
  rubric: string;
  subtype: string;
  topic_id: string;
  updated_at: string;
  users_to_evaluate: number;
  video: object;
}